+++ src/css.rs
@@ -4,14 +4,13 @@ use std::ops::Deref;
 
 use cssparser::{self, ToCss};
 use html5ever::LocalName;
-use tendril::StrTendril;
 
-/// CssString wraps StrTendril, to implement <ToCss> trait. Currently it is used in the <InnerSelector>.
+/// CssString wraps String, to implement `ToCss` trait. Currently it is used in the `InnerSelector`.
 #[derive(Clone, Eq, PartialEq, Debug)]
-pub struct CssString(StrTendril);
+pub struct CssString(String);
 
 impl Deref for CssString {
-    type Target = StrTendril;
+    type Target = String;
 
     fn deref(&self) -> &Self::Target {
         &self.0
@@ -26,7 +25,7 @@ impl AsRef<str> for CssString {
 
 impl From<&str> for CssString {
     fn from(value: &str) -> Self {
-        CssString(StrTendril::from(value))
+        CssString(String::from(value))
     }
 }