+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - expose `dom_tree::NodeData`.
 - minor adjustments.
 - revise `matcher::Matcher` usage, now `selection::Selection`'s underlying `select` methods uses `matcher::Matcher` reference if it possible.
+- `css::CssString` now wraps `String` instead of `StrTendril`. Because `StrTendril` can't be shared between multiple threads without mutex.
 
 ## [0.3.1] - 2023-12-31